fix: Changed path acquisition method and added prompt manual update operation
This commit updates the `run.sh` shell script to improve the handling of
the AWS profile in the shell prompt. The script uses the `BASH_SOURCE`
variable to get the full path to the script, and also adds a new
function called `update_ps1` to update the prompt with the AWS profile.
The `update_ps1` function uses regular expressions and ANSI escape codes
to replace the old profile in the `PS1` variable with the new profile.